ozone_automation_tool/requirements.txt
pynput
tkinter
re
mouse
Pillow==10.0.1
pytesseract==0.3.10